Splitting inferences for datatypes
